+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.0.3 - Moderation
+
+* Added moderation check
+* Removed new lines mid message
+
 ## 0.0.2 - Feature update
 
 * Added some missing features